bigmachine: tee and buffer output
Summary:
When a file descriptor on a machine is tailed (via Supervisor.Tail),
the implementation simply duplicated the file descriptor to a pipe
and wrote it to the receiver. This has many drawbacks: (1) it admits
only one tailed; (2) it doesn't perform proper tear down, effectively
limiting machines to one exec(); (3) without buffering, log output
could block a process and slow it down.

This change introduces a tee abstraction to manage buffered I/O
teeing to multiple destinations. Tee buffers output, per destination,
up to a configurable limit, and drops writes while the buffer remains
above the high watermark. Thus it's useful for logging output where
completeness is not critical for correctness.

Reviewers: pgopal, joshnewman, jheart

Reviewed By: jheart

Differential Revision: https://phabricator.grailbio.com/D12072

fbshipit-source-id: 392bb99