bigmachine: refactor API to permit easy testing; general simplifications
Summary:
This change introduces many API changes to Bigmachine that allow it
to be used in in-process testing contexts. Along with it comes a
number of API simplifications.

An obstacle to performing in-process testing was the fact that
services were registered globally. To overcome this, we change the
model slightly (and, I think, for the better): instead of registering
services, one or more service instances are presented when a new
machine is started. Because we do not know a-priori the possible set
of data types for services, this requires that any type that may
appear as a service must be registered with gob.Register. In
practice, I don't think this is a big problem.

We also simplify service invocation as part of these API changes:
instead of relying on service objects that name a service, machines
are now created with explicitly named services. These service names
are then provided directly ("ServiceName.MethodName") when making RPC
calls.

We provide a system implementation called testsystem that can be
used to test (in-process) bigmachine-based packages.

Reviewers: pgopal, jheart, joshnewman

Reviewed By: pgopal

Differential Revision: https://phabricator.grailbio.com/D12243

fbshipit-source-id: 937c41f