bigmachine: push log-tailing into system implemenations
Summary:
Log-tailing should be performed by the system: ssh access isn't
always possible (local), and the systems implementation should
encapsulate the specifics of how processes are bootrapped (e.g., the
fact that systemd is used, which unit, the username, ssh keys, etc.).

This change thus pushes log-tailing into the system implementation:

- bigmachine.Local is very simple: we just mirror (with a prefix) the
standard output and error from the cmd.

- ec2system: generate an ssh key for each system invocation; this key
is installed on the booted machines, and we use this key with Go's
ssh library to invoke journalctl to tail bootmachine's output.

Reviewers: ysaito, pgopal

Reviewed By: ysaito

Differential Revision: https://phabricator.grailbio.com/D14790

fbshipit-source-id: ed766ca