bigmachine/ec2system: retrieve root device from AMIs
Summary:
Bigmachine hardcoded the root image device, but not all AMIs follow
the same conventions (e.g. Ubuntu). This would cause us to allocate
an orphan root device and for the machine to be allocated with the
default root device specified in the AMI, usually causing the root
device to be too small.

Solve this by looking up AMIs and retrieving the root device from there.

Reviewers: ysaito, pgopal

Reviewed By: ysaito

Differential Revision: https://phabricator.grailbio.com/D25207

fbshipit-source-id: 6c93bf5