diviner: implement Bayesian optimization oracle through skopt
Summary:
This change implements a Bayesian optimization oracle using
scikit-optimize. This is done by constructing a Python script that
contains trial data, calling into the scikit optimizer, and then
printing out the new trial parameters as TSV which we then parse and
return as diviner values.

We also implement the ability run multiple rounds of optimizations
from diviner, so that, e.g., "diviner foo.dv run -parallelism 200
-trials 5000" will run 25 rounds of trials; each building a new model
which is used to sample parameter values for the next round. This
seems like the simplest strategy for parallel optimization, but I
expect we'll refine this in the future.

Reviewers: ysaito, kdavydov

Reviewed By: ysaito

Subscribers: onikolic, dfilippova, vnicula

Differential Revision: https://phabricator.grailbio.com/D25293

fbshipit-source-id: 7cbd5c9