bigmachine: make log-tailing part of the system implementation
Summary:
Log-tailing was done internally by each system, writing to the local
stdio. This causes a few problems, chiefly that it's difficult to tie
machine lifetimes to tailing lifetimes, so if you have a job with a
lot of machine churn (e.g., multiple diviner rounds), bigmachine
attempts to tail the stdio of each machine in perpetuity, and your
logs simply fill with tail failures.

This change remedies the situation by promoting log tailing to the
system level. Log tailing is cancelable and tied to the lifetime of
the machine. The way it's implemented also opens up future
improvements like giving users control over what output is tailed
locally as well as displaying logs content in web consoles, or
forwarding them to services like cloudwatch logs.

Reviewers: ysaito, pgopal

Reviewed By: ysaito

Differential Revision: https://phabricator.grailbio.com/D26511

fbshipit-source-id: 8d4f223