diviner/database: "value-oriented" runs, simplify
Summary:
This change changes the definition of Run from an interface to a
simple struct. Instead of definition operations directly on Run, they
are made via the Database interface.

This has a few advantages:

- it's clearer what's a database operation and what's not;
- we push keepalives/etc, to the runner; unifying and simplifying
  the database implementations;
- it's much easier to extend run in the future (e.g., to accomodate for
  datasets, or to add "properties");
- structs are generally easier to work with than interfaces.

We reap some immediate simplifications in the code. For example,
instead of complicated machinery to ensure that updates are rate
limited, and separately (asynchronous) keepalives, we instead simply
update each run every 15s with the latest status.

We also gain some functionality right away: we can now report all
metrics for a run, and we add the command "diviner metrics" to report
them all to be used for further analysis.

Along the way, I took the opportunity to clean up, generalize, and
simplify a few unrelated things. We also introduce package
"dynamoattr" that generalizes Go value marshaling and unmarshaling to
DynamoDB items. This may be useful elsewhere.

Reviewers: ysaito, joshnewman

Reviewed By: ysaito

Subscribers: ayip, vnicula, onikolic, dfilippova

Differential Revision: https://phabricator.grailbio.com/D27570

fbshipit-source-id: b96327c