bigmachine/client: Do not reset client on method errors
Summary: This revision eliminates the calls to `resetClient` when we encounter `methodErrorCode` errors.  This status code is returned when there is an error or panic executing the function on the remote machine, which means we can infer that our transport layer worked fine.  Resetting the client is an attempt to fix the transport layer (e.g. hung TCP), so we shouldn't need to do it in this scenario.

Test Plan:
Test with `bigslice` program with OOMing worker -- see many fewer connection resets; other failures still recover as expected (e.g. errors calling machines while they are booting).

@joshnewman tested with OOMing job; also showed reasonable behavior.

Reviewers: pgopal, smahadevan, marius, ysaito

Reviewed By: marius, ysaito

Subscribers: qliu, jschellenberger, joshnewman

Differential Revision: https://phabricator.grailbio.com/D28815

fbshipit-source-id: b3a3a85