bigmachine/ec2system: add rate limiting for requesting new instances
Summary:
Anecdotally, it appears that large Bigslice and Diviner jobs can get
into "thundering herds" when a large number of requests for machines
cause bursts of API requests; these in turn all use the same
(non-jittered) retry policy, and thus get in to a cycle of call
bursts, all of which cause a large number of RequestLimitExceeded
errors.

This fixes this by adding a rate limit as a gate to making EC2 API
requests (at least for machine allocation). The rate limit is global
to the process, and thus avoids the thundering herd issues described
above. We could also have added a jittering retry policy, but in this
case, the external limits are in fact rate limits (rather
arbitrarily), and so it seems slightly easier to reason about (and
possibly tune in the future).

This fixes the problem for some large Bigslice jobs I am currently
managing.

Reviewers: ysaito, joshnewman, jcharumilind, pgopal

Reviewed By: joshnewman

Subscribers: dfilippova

Differential Revision: https://phabricator.grailbio.com/D29175

fbshipit-source-id: 1a6be20