bigmachine: always use a recent timestamp for "ValidUntil"
Summary:
When retrying spot requests, ec2machine reused the "ValidUntil"
timestamp (the parameters were computed just once, and the same
request was re-issued). This caused many errors like

	2019/07/01 21:43:39 error starting machines: InvalidParameterValue: Value (2019-07-01T21:31:13 UTC) for parameter validUntil is invalid. The 'validUntil' time for this request has already expired.
	        status code: 400, request id: cdb8ea58-e6ae-41de-aa3d-f938456af0c6

Which slowed machine allocation and causes a lot of log spamming. (In
Bigslice, it means it got kicked to the outer retry loop.)

Reviewers: jcharumilind, ysaito, pgopal

Reviewed By: jcharumilind

Differential Revision: https://phabricator.grailbio.com/D32918

fbshipit-source-id: b55d520