base/dump: Add dump package for dumping diagnostics
Summary:
This revision adds a library that can be used to expose an endpoint that provides a tarball containing diagnostic files, some of which are provided by default, others of which can be registered by client code.

This revision also adds usage of this library to `slicecmd` (`/debug/dump`).

Test Plan: Unit tests; manual testing

Reviewers: pgopal, marius

Reviewed By: marius

Differential Revision: https://phabricator.grailbio.com/D32233

fbshipit-source-id: 3f8248c