separate binary upload from exec
This change separates binary upload from binary exec, so that a separate
timeout may be specified in teh case of the former. This allows
Bigmachine-systems to run in more bandwidth constrained scenarios.

This timeout is now computed based on a bandwidth floor of 100 kB/s.
(Arbitrary but reasonable.)

The change also adds a Getbinary method to the supervisor so that we can
distribute binaries among workers instead of re-uploading the binary
every time. (Implementing this is a future change.)

While this change improves the chances of success in bandwidth
constrained setups, using an overall timeout for stall detection is
still suboptimal. Future work includes monitoring upload progress
directly, and failing on stalls.