ec2system: add URL prefix to prevent address recycling
As outlined in #15, ec2system can currently recycle machine
addresses. This could be bad for downstream systems that depend on
addresses being unique (e.g., Bigslice storing the location of tasks
this way).

This fixes the problem by using the instance id as the URL prefix.
This has several advantages: first, it can be computed independently
in the driver and the worker; second, the URL carries meaningful
information: e.g., the EC2 instance ID is unique and can be used to
look up entries in cloudwatch, etc.