Close all connections when resetting clients (#35)
Close all connections when "resetting" clients. We reset clients by disposing of the clients and backing transports on error. Transports are only garbage collected when all connections, including idle connections, are closed. The existing code was not handling the case in which successful requests followed failed requests. We would eject the client from our cache, but subsequent successful requests would not close the remaining transport connections. The end result is that client resets could cause transport leaks. This is problematic for very long-running computations.

Change this to close idle connections on clients/transports that have been reset.