Prevent reboot of EC2 instances (#42)
Prevent reboot of EC2 instances. There are a few scenarios in which the instance may try to reboot:
- EC2 rebooting when encountering underlying hardware issues.
- EC2 scheduled maintenance.
- Services running that cause reboot, e.g. locksmithd. (This one is avoidable by using an appropriate AMI).

We prevent rebooting because bringing the instance back to a working state would add a lot of complexity, as we would need restore both internal state and the state of defined services.