Bookkeep muxer correctly for local system Tail (#45)
Bookkeep the muxer used to merge `Stdout` and `Stderr` from the machine process, which is used to tail logs from the machine. We construct a map keyed by machine pointer key, but we use a `nil` key. Fix to use the correct machine pointer.

`Tail` now works correctly, instead of showing the message, "machine not under management".