Fast fail on (service).Init panic (#64)
Fail fast when `(service).Init` panics. We make the assumption that a panic in `Init` is unrecoverable, so we fail fast, so it's faster (fails faster) and easier (less log noise) to recognize.

Note that we're taking a conservative approach. If `Init` merely returns a non-fatal error, we will still retry.