bigmachine: allow overriding remote bootstrap loader
Summary:
For experimental purposes:
  * I'm dynamically loading a Go-derived .so in a Python process (allowing Python-Go code sharing, etc.).
  * I want to run bigslice jobs, but bigmachine's view of "self" won't work (either Python binary or non-executable .so).
  * Building a directly executable .so with the Go toolchain doesn't seem obvious (as discussed offline).

With this change, I can build a separate remote bootstrap executable and embed it in the .so.

I'm happy to discuss alternatives if you see a better path.

Test Plan:
Existing bigmachine and bigslice tests exercise the default behavior.

Ran a bigslice job from a Jupyter notebook (the environment described above).

Reviewers: jcharumilind

Reviewed By: jcharumilind

Differential Revision: https://phabricator.grailbio.com/D55072

fbshipit-source-id: 679d54e