bigslice: schedule tasks even if highest priority task cannot be scheduled
Summary:
[BXDS-372]

Currently the scheduler will attempt to schedule the highest priority item.  If this fails, the scheduler blocks all other tasks from starting until the highest priority item can be scheduled.  This is wasteful when the highest priority item request many processors on a machine.  It can take a long time for a machine to become available and in the mean time, all other machines have no new work scheduled.  This causes the throughput of the cluster to suffer.  The new scheduler will look ahead to lower priority tasks while reserving the least occupied machine(s) for higher priority tasks that cannot currently be scheduled.  In practice on a workload with a mix of exclusive and small tasks, this can lead to more efficient resource utilization and faster execution.

Here is an example trace using the existing scheduler.  Notice the pauses in scheduling while blocked on the large top-priority task that is ultimately scheduled on a different machine.
{F1397276}

Here is a similar section using the scheduler implemented in this commit.  Notice the lack of any pausing.
{F1397278}

Along the way, I made the tests clean up their `testsystem` servers to avoid running out of sockets when doing some stress testing.

Co-authored with @jschellenberger; see D57990

Test Plan: Run (added) unit tests successfully.  Run `classifier-evalution` in weekly configuration successfully.

Reviewers: jschellenberger, joshnewman

Reviewed By: jschellenberger, joshnewman

Subscribers: jschellenberger

Differential Revision: https://phabricator.grailbio.com/D59255

fbshipit-source-id: ec1d232