bigmachine: kill redundant printing of machine addr
Summary:
Kill the redundant printing of the local machine address to get rid of redundancy in logs.

Logs from machines are already prefixed with the machine address by the driver, so having the tail implementation also add it leads to redundancy, e.g.:

```
https://localhost:44201/: localhost:44201: 2021/06/03 23:23:18 610000000/2000000000
```

Notice the double printing of the `host:port`.

Test Plan: Run `bigpi` and see no more redundancy.

Reviewers: joshnewman

Reviewed By: joshnewman

Differential Revision: https://phabricator.grailbio.com/D60012

fbshipit-source-id: 24aac73