[BXDS-358] chiron/tfexec: invoke training with tfexec in classifier-evaluation process
Summary:
Invoke training with `tfexec` in the classifier-evaluation process instead of starting separate processes for each training.  This will make it easier to implement more sophisticated scheduling of trainings on machines, e.g. reusing machines, without the extra complexity of inter-process communication.

This is now possible because of the recently added ability to have multiple bigmachine instances within the same process.

This does not implement a more sophisticated scheduler.  A machine is still started for each training.  The updated scheduler will come later.

Test Plan: Run unit tests.  Run sexandage with `tfexec` in classifier-evaluation, and see that it still works.

Reviewers: O52 chiron infrastructure, joshnewman

Reviewed By: O52 chiron infrastructure, joshnewman

Subscribers: anandi, dborcherding

Differential Revision: https://phabricator.grailbio.com/D61185

fbshipit-source-id: dbee69f