[BXDS-736] bigmachine/ec2system: wait until instance exists before tagging
Summary: Wait until instances exist before we attempt to tag them.  We have recently seen what appears to be a race in which the attempt to tag the launched instance fails because the instance ID is unknown (due to a consistency lag).  Explicitly wait so our attempt to tag is more likely to succeed.

Test Plan: Run `bigslice-demo`.  See that all instances are tagged.

Reviewers: joshnewman

Reviewed By: joshnewman

Differential Revision: https://phabricator.grailbio.com/D64198

fbshipit-source-id: fa5d062