[BXDS-784] bigmachine/ec2system: error if waiting for running state fails
Summary: Return error if waiting for running state fails.  When we launch instances, we call `WaitUntilInstanceRunningWithContext` before we retrieve details about our launched instances.  When `WaitUntilInstanceRunningWithContext` returns an error, we may end up returning and empty slice of machines and a `nil` error (I think because of previous oversight).  In all other cases in which `Start` returns an empty slice of machines, we report a non-`nil` error describing why we were unable to launch an instance/machine.  Make this scenario consistent, and return a non-`nil` error.

Test Plan: Run unit tests.  Run `bigslice-demo`, and see that it still works.

Reviewers: joshnewman

Reviewed By: joshnewman

Differential Revision: https://phabricator.grailbio.com/D64267

fbshipit-source-id: 1a098f7