[BXDS-1178]bigslice: fail fast if error gob-encoding func args
Summary: Fail fast if there is an error when gob-encoding the arguments to a `Func` invocation.  The existing code sees the problem on the attempt to `POST` to make the RPC.  The fatal severity of the error returned from the reader (providing the stream to be sent to the server) is not propagated out of the `http` client, as that happens in standard library code.  The code in this commit adds the ability for the `io.Reader`-making function to return an error which is propagated before the attempt to `POST`.

Test Plan: Run added unit test.  Run the case that motivated the change, D68357, and see fast failure.

Reviewers: joshnewman

Reviewed By: joshnewman

Subscribers: fchu

Differential Revision: https://phabricator.grailbio.com/D68419

fbshipit-source-id: 41016b2